refactor(web): extract useMessageDispatch from useChatWithSync
Move the most coupled slice of the chat hook into useMessageDispatch:
sendMessage (draft materialization, optimistic update, pull-conflict +
daily-limit handling, stream kickoff), stopAgent, the per-chat send
queue (useQueueDispatch), and the sendInFlight/stopInFlight refs. Its
dependency surface is an explicit args object.

Public return shape unchanged. Behavior-preserving except sendMessage's
deps array drops two entries unused in its body (previewStates,
updateChatById) which only caused needless recreation.

Completes the useChatWithSync split: 833 -> 385 lines across 4 hooks
(useChatWithSync, useMessageDispatch, useChatOperations,
useChatMessageSync).
